A discount percentage is a fraction, like the tax rate
The till's `DiscountKind::Percentage` value was percent points (`10` =
10%) and the engine divided by 100. The tax rate beside it was already a
fraction that the engine multiplied by. Two conventions for "a
percentage" in one pricing function, disambiguated only by which field
you were reading.

`price_cart` now multiplies by the fraction — the same arithmetic, and
the same half-away-from-zero rounding, as the backend's `calc_discount`.
The hand-written reference oracle is restated to match, so cargo-mutants
still cannot mutate the engine and its check together, and the proptest
strategy keeps both regions: fractions past 1.0 to exercise the >100%
clamp, and a wider range so a Fixed discount is still realistic.

`i64` became `f64` all the way out to the FFI, because the point of the
change is that 12.5% is now expressible at all — an integer percentage
could not say it, while a 12.5% tax rate was ordinary.

The one place a customer sees it — the discount chip in the checkout
drawer — multiplies back up, and drops a trailing `.0` so the common
case still reads "10%".

+++ rust-core/crates/madar-api/src/models/create_discount_request.rs
@@ -34,7 +34,7 @@ pub struct CreateDiscountRequest {
     #[serde(rename = "org_id")]
     pub org_id: uuid::Uuid,
     #[serde(rename = "value")]
-    pub value: i32,
+    pub value: f64,
 }
 
 impl CreateDiscountRequest {
@@ -42,7 +42,7 @@ impl CreateDiscountRequest {
         dtype: String,
         name: String,
         org_id: uuid::Uuid,
-        value: i32,
+        value: f64,
     ) -> CreateDiscountRequest {
         CreateDiscountRequest {
             dtype,
