+++ rust-core/crates/madar-api/src/models/create_open_ticket_request.rs
@@ -51,7 +51,7 @@ pub struct CreateOpenTicketRequest {
         with = "::serde_with::rust::double_option",
         skip_serializing_if = "Option::is_none"
     )]
-    pub discount_value: Option<Option<i32>>,
+    pub discount_value: Option<Option<f64>>,
     #[serde(
         rename = "guest_count",
         default,
